Update SellerDTO.java
sellerDTO int member_id 수정

+++ src/main/java/com/pt/brewit/dto/SellerDTO.java
@@ -6,7 +6,7 @@ import java.util.Date;
 @Data
 public class SellerDTO {
     private int seller_id;        // seller_id
-    private long member_id;       // member_id
+    private int member_id;       // member_id
     private String company_name;   // company_name
     private String b_number;      // b_number
     private Date reg_date;        // reg_date
